requirements.txt
streamlit>=1.31.0
opencv-python-headless>=4.8.1
deepface>=0.0.79
numpy>=1.24.3
Pillow>=10.0.0
pandas>=2.0.3
